# Conflicts:
#	app/src/main/java/com/marluki/misterymap/MainActivity.java

+++ app/src/main/java/com/marluki/misterymap/sync/SyncAdapter.java
@@ -244,8 +244,8 @@ public class SyncAdapter extends AbstractThreadedSyncAdapter implements GoogleAp
 
                     id = c.getString(c.getColumnIndex(DatuBaseKontratua.Objetos_mapa.ID));
                     ObjetoMapa objetoMapa = new ObjetoMapa(id, c.getInt(c.getColumnIndex(DatuBaseKontratua.Objetos_mapa.TIPO_ID)),
-                            c.getFloat(c.getColumnIndex(DatuBaseKontratua.Objetos_mapa.LATITUD)),
-                            c.getFloat(c.getColumnIndex(DatuBaseKontratua.Objetos_mapa.LONGITUD)),
+                            c.getDouble(c.getColumnIndex(DatuBaseKontratua.Objetos_mapa.LATITUD)),
+                            c.getDouble(c.getColumnIndex(DatuBaseKontratua.Objetos_mapa.LONGITUD)),
                             c.getString(c.getColumnIndex(DatuBaseKontratua.Objetos_mapa.USUARIO_ID)),
                             c.getString(c.getColumnIndex(DatuBaseKontratua.Objetos_mapa.NOMBRE_OBJETO)),
                             c.getString(c.getColumnIndex(DatuBaseKontratua.Objetos_mapa.DETALLES)),
